Develop

+++ src/main/java/com/example/demo/customer/controller/CustomerController.java
@@ -24,7 +24,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}")
-    public CustomerDto getCustomer(@PathVariable Long id) {
+    public CustomerDto getCustomer(@PathVariable Integer id) {
         return customerService.getCustomerById(id);
     }
 
@@ -35,24 +35,24 @@ public class CustomerController {
     }
 
     @PutMapping("/{id}")
-    public CustomerDto updateCustomer(@PathVariable Long id,
+    public CustomerDto updateCustomer(@PathVariable Integer id,
             @RequestBody CustomerUpdateDto dto) {
         return customerService.updateCustomer(id, dto);
     }
 
     @PutMapping("/{id}/address")
-    public AddressDto updateAddress(@PathVariable Long id,
+    public AddressDto updateAddress(@PathVariable Integer id,
             @RequestBody AddressDto dto) {
         return customerService.updateAddress(id, dto);
     }
 
     @GetMapping("/{id}/rentals")
-    public List<RentalResponseDto> getRentals(@PathVariable Long id) {
+    public List<RentalResponseDto> getRentals(@PathVariable Integer id) {
         return customerService.getCustomerRentals(id);
     }
 
     @GetMapping("/{id}/payments")
-    public List<PaymentDto> getPayments(@PathVariable Long id) {
+    public List<PaymentDto> getPayments(@PathVariable Integer id) {
         return customerService.getCustomerPayments(id);
     }
 
@@ -67,22 +67,22 @@ public class CustomerController {
     }
 
     @PatchMapping("/{id}")
-    public CustomerDto patchCustomer(@PathVariable Long id, @RequestBody CustomerUpdateDto dto) {
+    public CustomerDto patchCustomer(@PathVariable Integer id, @RequestBody CustomerUpdateDto dto) {
         return customerService.patchCustomer(id, dto);
     }
 
     @PatchMapping("/{id}/address")
-    public AddressDto patchAddress(@PathVariable Long id, @RequestBody AddressDto dto) {
+    public AddressDto patchAddress(@PathVariable Integer id, @RequestBody AddressDto dto) {
         return customerService.patchAddress(id, dto);
     }
 
     @GetMapping("/{id}/payment/{paymentId}")
-    public PaymentDto getPaymentById(@PathVariable Long id, @PathVariable Long paymentId) {
+    public PaymentDto getPaymentById(@PathVariable Integer id, @PathVariable Integer paymentId) {
         return customerService.getCustomerPaymentById(id, paymentId);
     }
 
     @GetMapping("/{id}/rentals/{rentalId}")
-    public RentalResponseDto getRentalById(@PathVariable Long id, @PathVariable Long rentalId) {
+    public RentalResponseDto getRentalById(@PathVariable Integer id, @PathVariable Integer rentalId) {
         return customerService.getCustomerRentalById(id, rentalId);
     }
 }
