+++ src/main/java/com/example/demo/customer/repository/PaymentRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
-    List<Payment> findByCustomerId(Long customerId);
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
+    List<Payment> findByCustomerId(Integer customerId);
 
     @Query("SELECT p FROM Payment p WHERE p.customerId = :customerId AND p.paymentId = :paymentId")
-    java.util.Optional<Payment> findByCustomerIdAndPaymentId(@org.springframework.data.repository.query.Param("customerId") Long customerId, @org.springframework.data.repository.query.Param("paymentId") Long paymentId);
+    java.util.Optional<Payment> findByCustomerIdAndPaymentId(@org.springframework.data.repository.query.Param("customerId") Integer customerId, @org.springframework.data.repository.query.Param("paymentId") Integer paymentId);
 }
